backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-jose
passlib[bcrypt]
pydantic
